pyproject.toml
[tool.poetry]
name = "paid-content"
version = "0.1.0"
description = ""
authors = ["Nikita-Musk <nik52119w@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
tree = "^0.2.4"
ipython = "^8.29.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
stripe = "^11.2.0"
coverage = "^7.6.7"
twilio = "^9.3.7"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
